GenAI-Projects/week1_finetuning_llm/requirements.txt
openai>=1.3.7
pandas>=1.5.0
numpy>=1.23.0
tqdm>=4.64.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
requests>=2.28.0